Add solution for May 24th, 2024 daily challenge
A new function, maxScoreWords1, was added for the daily challenge of May 24th, 2024. This function finds the maximum score of any valid set of words that can be formed using the given letters, using a backtracking approach. This solution's time complexity is O(2^n * L), where n is the number of words, and L is the average length of the words in the input list.