Add detailed answer explanation in May20-26 notebook
In this commit, a detailed explanation of a problem solution has been added to the May20-26 notebook. It covers the process of initialization, backtracking, finding overall maximum, and the role of memoization. It also corresponds to changes in the "ExecuteTime" metadata to "is_executing" and sets "execution_count" as null.