Refactor and complete attendance record calculation functions
The definitions and comments in these functions have been updated to be more clear and concise. Also, the second function, `checkRecord2()`, which formerly lacked implementation, has been filled with an iterative dynamic programming solution to calculate the number of valid attendance records. This change allows a performance comparison between the top-down and bottom-up dynamic programming approaches for this problem.