Remove TODO and add equalSubstring function
In the May20-26.py file, removed the TODO comments. In the May27-31.py, a new function 'equalSubstring' is added which computes the length of the longest substring that can be transformed from one string to another within a given budget cost. This function uses a sliding window approach. Also, improvements have been made to existing comments and test cases are provided for the new function.