Implement and test 'appendCharacters1' function
The function 'appendCharacters1' has been implemented. This function determines the minimum number of characters from a given string 't' that need to be appended to another string 's' to make 't' a subsequence of 's'. The function was also tested for accuracy. A detailed explanation of the algorithm and its time and space complexities are included in the function's docstring for future reference.