Implement common character finder function
A function called "commonChars1" has been created. This function accepts a list of words and returns an array of all characters that appear in all strings within the list, including duplicates. The time and space complexity of the function has also been analyzed and outlined in the function docstring. Additionally, test cases have been added for verification.