Implement function to find common characters in strings
The function `commonChars2` was implemented to find common characters (including duplicates) in all strings that are presented in the input list `words`. A bitwise 'AND' operation is performed on a 'Counter' object of each word, resulting in the intersection of characters. Also, the test calls for the `commonChars1` and `commonChars2` functions were commented out.