Update sentence parsing in algorithm explanation
An explanation of how we were parsing sentences in the algorithm has been clarified. To be more precise, we're creating substrings from the minimum length of dictionary words to the minimum length between the maximum length of dictionary words and the length of the current word. The overall goal is to improve our comprehension of the strategy.