Add heightChecker2 function with counting sort algorithm
The heightChecker2 function has been implemented which counts the number of students not standing in their expected height order using the Counting Sort algorithm. The function works by creating a deepcopy of the heights list, applying counting sort, and then comparing elements in both lists. Test cases have also been updated to include checks for this new function. The time and space complexity of this solution is O(n + k), with 'n' being the number of elements in the heights' list and 'k' being the range of values.