Implement sorting methods in daily question notebook
Two different sorting methods are implemented in the Jun10-16.ipynb daily question notebook. The methods are used to sort an array such that the first part of the result has elements in the same relative order as another given array. The first approach uses a hash map and sorting, while the second uses counting sort.