Implement minMovesToSeat1 function
This commit includes the implementation of the function minMovesToSeat1. This function calculates the minimum number of moves to seat everyone. It sorts both the given lists, loops through them and adds the absolute difference of the students and seats to the total moves. A test case for this function is also added.