Implement function to solve 'Sum of Square Numbers' problem
Updated the Python and Jupyter notebook files for the week of June 17-23, 2024 to include the 'Sum of Square Numbers' problem. In the Python code, added the function judgeSquareSum1 to solve the problem, while in the Jupyter notebook, added the problem statement, examples, and constraints of the 'Sum of Square Numbers' problem scheduled for June 17th, 2024.