Implement function for minimum bloom days
The function `minDays1` for calculating minimum bloom days has been implemented. The function checks if making the desired number of bouquets is possible within the given number of days and uses a binary search algorithm to find the minimum days it takes for the flowers to bloom. Changes in test cases and notebook annotations were made to reflect this.