Implement maxBoxesInWarehouse2 function
The function maxBoxesInWarehouse2 was previously a placeholder. This commit populates it with an implementation that calculates the maximum number of boxes that can be accommodated in the warehouse, given the box and warehouse dimensions. The algorithm is designed to be efficient by sorting the boxes and utilizing two pointers to traverse the warehouse list.