Add new topics for 2024 exercise compilations and weekly programming exercises
Several new markdown files have been added for compiling daily and weekly exercises for the year 2024. These include monthly files for each month of the year, as well as a general template for daily and weekly programming exercises. These files will serve as the structure for future exercise compilations and enable easy navigation for users.